Add files via upload
Here we gonna learn how the while loop operates in javascript and what is the major difference between for loop and while loop.One major use case of while loop is that we would be able to keep on printing the randomly. Suppose we want to roll a dice until to the point when its not come to six. Trunc method is basically used to avoid the decimal part while generating the random numbers between any two value.